Move Editor files to src folder
