+++ src/Place/ReadModel/RDF/RdfProjector.php
@@ -26,10 +26,10 @@ use CultuurNet\UDB3\Place\Events\Moderation\Rejected;
 use CultuurNet\UDB3\Place\Events\PlaceDeleted;
 use CultuurNet\UDB3\Place\Events\TitleTranslated;
 use CultuurNet\UDB3\Place\Events\TitleUpdated;
-use CultuurNet\UDB3\RDF\GraphEditor;
+use CultuurNet\UDB3\RDF\Editor\GraphEditor;
 use CultuurNet\UDB3\RDF\GraphRepository;
 use CultuurNet\UDB3\RDF\MainLanguageRepository;
-use CultuurNet\UDB3\RDF\WorkflowEditor;
+use CultuurNet\UDB3\RDF\Editor\WorkflowEditor;
 use EasyRdf\Graph;
 use EasyRdf\Literal;
 use EasyRdf\Resource;
